put tester output in external files
