+++ test/data/example-digest-bad.warc.test
@@ -1,0 +1,22 @@
+test/data/example-digest-bad.warc
+  WARC-Record-ID <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+    WARC-Type request
+    payload digest failed: sha1:1112H3S6NNFQ2MSVX7XZKYAYSCX5QBYJ
+    error: WARC-IP-Address should be used for http and https requests
+  WARC-Record-ID <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+    WARC-Type request
+    digest pass
+    error: WARC-IP-Address should be used for http and https requests
+    error: Duplicate WARC-Record-ID: <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+  WARC-Record-ID <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+    WARC-Type request
+    digest pass
+    error: WARC-IP-Address should be used for http and https requests
+    error: Duplicate WARC-Record-ID: <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+  WARC-Record-ID <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+    WARC-Type request
+    digest pass
+    error: WARC-IP-Address should be used for http and https requests
+    error: Duplicate WARC-Record-ID: <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+global Concurrent-To checks
+  comment: WARC-Concurrent-To not found: <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007> WARC-Concurrent-To <urn:uuid:a9c51e3e-0221-11e7-bf66-0242ac120005>
