+++ test/data/example.warc.test
@@ -1,0 +1,16 @@
+test/data/example.warc
+  WARC-Record-ID <urn:uuid:a9c5c23a-0221-11e7-8fe3-0242ac120007>
+    WARC-Type request
+    digest not present
+    error: WARC-IP-Address should be used for http and https requests
+  WARC-Record-ID <urn:uuid:e6e395ca-0221-11e7-a18d-0242ac120005>
+    WARC-Type revisit
+    digest present but not checked (revisit)
+    recommendation: Missing recommended header: WARC-Refers-To
+    comment: This Heretrix extension never made it into the standard: WARC-Profile http://netpreserve.org/warc/1.0/revisit/uri-agnostic-identical-payload-digest
+    comment: Field was introduced after this warc version: 1.0 WARC-Refers-To-Target-URI http://example.com/
+    comment: Field was introduced after this warc version: 1.0 WARC-Refers-To-Date 2017-03-06T04:02:06Z
+  WARC-Record-ID <urn:uuid:e6e41fea-0221-11e7-8fe3-0242ac120007>
+    WARC-Type request
+    digest not present
+    error: WARC-IP-Address should be used for http and https requests
